refactor(engine): give removeGlobalConfig its own module
Pure move out of pipeline.ts plus a cached globalOnlyOptionNames(), so
presentation layers can prune the 107 globalOnly options without
reaching into the pipeline. Call sites unchanged; golden and shimmed
suites prove the move is behavior-free.
